cleanup and colours
