+++ software/drivers/common.py
@@ -1,7 +1,18 @@
+from log import colours
+
 # Commands:
 PING = b"\x00"
 SENSORS_ALL = b"\x30"
 SENSOR_ACCELEROMETER = b"\x31"
+
+def deadzone(value, radius):
+	if abs(value) < 10:
+		value = 0
+	elif value > 0:
+		value -= radius
+	elif value < 0:
+		value += radius
+	return value
 
 class Sensors:
     def __init__(self, accel, gyro, flex):
@@ -10,7 +21,7 @@ class Sensors:
         self.flex = flex
 
     def __str__(self):
-        return str(self.accel) + str(self.gyro) + str(self.flex)
+        return f"{str(self.accel)}\n{str(self.gyro)}\n{str(self.flex)}\n"
 
 class Accelerometer:
     def __init__(self, x=0, y=0, z=0):
@@ -19,7 +30,11 @@ class Accelerometer:
         self.z = z
 
     def __str__(self):
-        return "| X: %6.0f | Y: %6.0f | Z: %6.0f |\n" % (self.x, self.y, self.z)
+        return "| %sX%s: %6.0f | %sY%s: %6.0f | %sZ%s: %6.0f |" % (
+            colours.BLUE, colours.RESET, self.x,
+            colours.BLUE, colours.RESET, self.y,
+            colours.BLUE, colours.RESET, self.z
+        )
 
 class Gyro:
     def __init__(self, pitch=0, yaw=0, roll=0):
@@ -28,7 +43,11 @@ class Gyro:
         self.yaw = yaw
 
     def __str__(self):
-        return "| PITCH: %6.0f | ROLL: %6.0f | YAW: %6.0f |\n" % (self.pitch, self.roll, self.yaw)
+        return "| %sPITCH%s: %6.0f | %sROLL%s: %6.0f | %sYAW%s: %6.0f |" % (
+            colours.RED, colours.RESET, self.pitch,
+            colours.RED, colours.RESET, self.roll,
+            colours.RED, colours.RESET, self.yaw
+        )
 
 
 class Flex:
@@ -40,4 +59,10 @@ class Flex:
         self.pinky = pinky
 
     def __str__(self):
-        return "| THUMB: %4.0f | INDEX: %4.0f | MIDDLE: %4.0f | RING: %4.0f | PINKY: %4.0f |\n" % (self.thumb, self.index, self.middle, self.ring, self.pinky)
+        return "| %sTHUMB%s: %4.0f | %sINDEX%s: %4.0f | %sMIDDLE%s: %4.0f | %sRING%s: %4.0f | %sPINKY%s: %4.0f |" % (
+            colours.YELLOW, colours.RESET, self.thumb,
+            colours.YELLOW, colours.RESET, self.index,
+            colours.YELLOW, colours.RESET, self.middle,
+            colours.YELLOW, colours.RESET, self.ring,
+            colours.YELLOW, colours.RESET, self.pinky
+        )
